Ster 64 rc import template refactor (#961)
* Refactor to custom script use

* Refactor
