+++ tools/importer/about.js
@@ -11,6 +11,7 @@
  */
 /* global WebImporter */
 /* eslint-disable no-console, class-methods-use-this */
+
 const baseDomain = 'https://main--shredit--stericycle.aem.page';
 const req = new XMLHttpRequest();
 let tags = {};
@@ -166,19 +167,38 @@ function setMetadata(meta, document) {
   }
 }
 
+function decodeAndReplace(str) {
+  const replacements = {
+    '%C3%A1': 'a',
+    '%C3%A9': 'e',
+    '%C3%AD': 'i',
+    '%C3%B3': 'o',
+    '%C3%BA': 'u',
+    '%C3%B1': 'n',
+  };
+  let updatedStr = str;
+  Object.entries(replacements).forEach(([encoded, plain]) => {
+    const regex = new RegExp(encoded, 'g');
+    updatedStr = updatedStr.replace(regex, plain);
+  });
+  return updatedStr;
+}
+
+
 /**
  * Updates the given URL by replacing consecutive hyphens in the pathname with a single hyphen,
  * and removes a trailing hyphen from the pathname if present.
  *
- * @param {Object} params - The parameters object.
- * @param {string} params.originalURL - The original URL to be updated.
+ * @param {string} originalURL - The original URL to be updated.
  * @returns {string} The updated, valid URL as a string.
  */
-function updatetoValidUrl(params) {
-  const url = new URL(params.originalURL);
+function updatetoValidUrl(originalURL) {
+  const url = new URL(originalURL);
   url.pathname = url.pathname.replace(/--+/g, '-');
-  url.pathname = url.pathname.endsWith('-') ? url.pathname.slice(0, -1) : url.pathname;
-  return url.toString();
+  url.pathname = url.pathname.endsWith('-')
+    ? url.pathname.slice(0, -1)
+    : url.pathname;
+  return decodeAndReplace(url.toString());
 }
 
 export default {
@@ -248,7 +268,7 @@ export default {
     // eslint-disable-next-line no-unused-vars
     document, url, html, params,
   }) => {
-    url  = updatetoValidUrl(params);
+    url  = updatetoValidUrl(params.originalURL);
     let p = new URL(url).pathname;
     if (p.endsWith('/')) {
       p = `${p}index`;
