+++ tools/importer/blog.js
@@ -70,6 +70,39 @@ function fixDynamicMedia(main, document) {
   });
 }
 
+function decodeAndReplace(str) {
+  const replacements = {
+    '%C3%A1': 'a',
+    '%C3%A9': 'e',
+    '%C3%AD': 'i',
+    '%C3%B3': 'o',
+    '%C3%BA': 'u',
+    '%C3%B1': 'n',
+  };
+  let updatedStr = str;
+  Object.entries(replacements).forEach(([encoded, plain]) => {
+    const regex = new RegExp(encoded, 'g');
+    updatedStr = updatedStr.replace(regex, plain);
+  });
+  return updatedStr;
+}
+
+/**
+ * Updates the given URL by replacing consecutive hyphens in the pathname with a single hyphen,
+ * and removes a trailing hyphen from the pathname if present.
+ *
+ * @param {string} originalURL - The original URL to be updated.
+ * @returns {string} The updated, valid URL as a string.
+ */
+function updatetoValidUrl(originalURL) {
+  const url = new URL(originalURL);
+  url.pathname = url.pathname.replace(/--+/g, '-');
+  url.pathname = url.pathname.endsWith('-')
+    ? url.pathname.slice(0, -1)
+    : url.pathname;
+  return decodeAndReplace(url.toString());
+}
+
 export default {
   /**
      * Apply DOM operations to the provided document and return
@@ -136,6 +169,7 @@ export default {
     // eslint-disable-next-line no-unused-vars
     document, url, html, params,
   }) => {
+    url = updatetoValidUrl(url);
     let p = new URL(url).pathname;
     if (p.endsWith('/')) {
       p = `${p}index`;
